requirements.txt
discord.py
aiohttp>=3.9.0
python-dotenv
